Create PlotUtils, plot_averaged_genes
No doubt it will be helpful to have several plotting functions, which
I'm going to put in a separate module.  The first one is
plot_averaged_genes, which will stitch together the data from the
upstream region, the gene itself, and the downstream region, and then
relabel the axes so that everything is properly labelled.