Minor bug fixes
I had labelled the array for the gene itself as "gene", which then got
reused for the entries from the GTF data.  Now it's gene_cov.

I'm also making sure to only use entries from the bacterial chromosome,
not anything from the F plasmid, which is also in the GTF file.