Better ignore non-gene based expression
When calculating nearest_to_{left,right}, do it pretty conservatively,
so that almost every annotated thing is included.  On the other hand,
when calculating the gene coverage, only use CDS's.  Between when I
calculate the nearest genes and the coverages, I now need to filter out
everything that's not a CDS, and for that I need to a) retain the type,
and b) actually filter baed on that.