AD HOC FIX: overlapping annotations
There seem to be two annotations that overlap, so that my method of
looking for coverage doesn't quite work.  Given that there's only one
instance of overlap, I'm okay for now just having a temporary fix, but I
should probably put in a function just for checking that a given
position is clear of anything else.