Allow arbitrary scaling of coding region in plot_averaged_genes
Do this by use of a bp_equiv argument, which says that the CDS area
should be equivalent to that many basepairs of the upstream or
downstream region.

This also has the advantage of removing magic numbers from the body of
the function, which should help make the code more comprehensible to
others.