Add UTR Compare
This reads in the bigwig files for a drug and a no-drug case, and
compares the expression in the UTRs for those cases.  It store those
ratios into the global variable all_ratios, and will optionally output
the UTRs whose ratio falls within a given range.

This was built up from a very simple program, and had stuff bolted on,
refactored, etc.  Sadly, this was not committed at any point, since it
all happened pretty quickly.