Code for plotting the FFT of a sequence
FFTSeq doesn't actually do this. It simply takes a fasta file and turns
the given base pairs into a 1, leaving everything else a 0. This will
likely need to be modified for looking for di, tri, and other
oligo-nucleotide sequences.

The plotting actually happens in a new function in PlotUtils, which was
originally just me typing out a one liner and changing the appropriate
line, but this was determined to be too clunky and unreliable.